+++ ice-20201109/src/Models/SubmitScreenMediaHighlightsJobRequest.php
@@ -1,0 +1,92 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\ICE\V20201109\Models;
+
+use AlibabaCloud\Tea\Model;
+
+class SubmitScreenMediaHighlightsJobRequest extends Model
+{
+    /**
+     * @example {
+     * }
+     * @var string
+     */
+    public $editingConfig;
+
+    /**
+     * @example {
+     * "MediaArray": [
+     * "****9d46c886b45481030f6e****",
+     * "****6c886b4549d481030f6e****"
+     * }
+     * @var string
+     */
+    public $inputConfig;
+
+    /**
+     * @example {
+     * }
+     * @var string
+     */
+    public $outputConfig;
+
+    /**
+     * @var string
+     */
+    public $userData;
+    protected $_name = [
+        'editingConfig' => 'EditingConfig',
+        'inputConfig'   => 'InputConfig',
+        'outputConfig'  => 'OutputConfig',
+        'userData'      => 'UserData',
+    ];
+
+    public function validate()
+    {
+    }
+
+    public function toMap()
+    {
+        $res = [];
+        if (null !== $this->editingConfig) {
+            $res['EditingConfig'] = $this->editingConfig;
+        }
+        if (null !== $this->inputConfig) {
+            $res['InputConfig'] = $this->inputConfig;
+        }
+        if (null !== $this->outputConfig) {
+            $res['OutputConfig'] = $this->outputConfig;
+        }
+        if (null !== $this->userData) {
+            $res['UserData'] = $this->userData;
+        }
+
+        return $res;
+    }
+
+    /**
+     * @param array $map
+     *
+     * @return SubmitScreenMediaHighlightsJobRequest
+     */
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['EditingConfig'])) {
+            $model->editingConfig = $map['EditingConfig'];
+        }
+        if (isset($map['InputConfig'])) {
+            $model->inputConfig = $map['InputConfig'];
+        }
+        if (isset($map['OutputConfig'])) {
+            $model->outputConfig = $map['OutputConfig'];
+        }
+        if (isset($map['UserData'])) {
+            $model->userData = $map['UserData'];
+        }
+
+        return $model;
+    }
+}